master/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.0
httpx[http2]>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
geopy>=2.4.0